Revert "Merge pull request #1842 from helsinki-systems/fix/64bit-timestamps"
We we want to try getting back to Nix 2.35 (the latest release at this
time) before attempting the timestamp migration, which may be slow.

This reverts commit 172395f97e3f42e04fd50cfd9e5d98c42a6cf094, reversing
changes made to aba9daf2365598859bb9134d645acfebf6d7cf64.

+++ subprojects/crates/db/src/models.rs
@@ -115,8 +115,8 @@ impl Build<String> {
 
 #[derive(Debug, Clone, Copy)]
 pub struct BuildSteps {
-    pub starttime: Option<i64>,
-    pub stoptime: Option<i64>,
+    pub starttime: Option<i32>,
+    pub stoptime: Option<i32>,
 }
 
 #[repr(i32)]
@@ -129,8 +129,8 @@ pub enum BuildType {
 #[derive(Debug)]
 pub struct UpdateBuild<'a> {
     pub status: BuildStatus,
-    pub start_time: i64,
-    pub stop_time: i64,
+    pub start_time: i32,
+    pub stop_time: i32,
     pub size: i64,
     pub closure_size: i64,
     pub release_name: Option<&'a str>,
@@ -144,8 +144,8 @@ pub struct InsertBuildStep<'a> {
     pub drv_path: &'a StorePath,
     pub status: BuildStatus,
     pub busy: bool,
-    pub start_time: Option<i64>,
-    pub stop_time: Option<i64>,
+    pub start_time: Option<i32>,
+    pub stop_time: Option<i32>,
     pub platform: Option<&'a str>,
     pub propagated_from: Option<i32>,
     pub error_msg: Option<&'a str>,
@@ -162,7 +162,7 @@ pub struct InsertBuildStep<'a> {
 pub struct InsertResolvedBuildStep<'a> {
     pub build_id: BuildID,
     pub drv_path: &'a StorePath,
-    pub start_time: i64,
+    pub start_time: i32,
     pub platform: Option<&'a str>,
     pub machine: &'a str,
     pub resolved_drv_path: &'a StorePath,
@@ -203,8 +203,8 @@ pub struct UpdateBuildStepInFinish<'a> {
     pub step_nr: i32,
     pub status: BuildStatus,
     pub error_msg: Option<&'a str>,
-    pub start_time: i64,
-    pub stop_time: i64,
+    pub start_time: i32,
+    pub stop_time: i32,
     pub machine: Option<&'a str>,
     pub overhead: Option<i32>,
     pub times_built: Option<i32>,
@@ -233,7 +233,7 @@ pub(crate) struct InsertBuildMetric<'a> {
     pub project: &'a str,
     pub jobset: &'a str,
     pub job: &'a str,
-    pub timestamp: i64,
+    pub timestamp: i32,
 }
 
 #[derive(Debug)]
